+++ prj/vs2022/CvtOnnx.vcxproj.filters
@@ -153,6 +153,9 @@
     <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertInput.cpp">
       <Filter>OnnxRuntime\I</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertInstanceNormalizationNode.cpp">
+      <Filter>OnnxRuntime\I</Filter>
+    </ClCompile>
     <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertConcatNode.cpp">
       <Filter>OnnxRuntime\C</Filter>
     </ClCompile>
